added casting to remove compiler errors
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@21933 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Filters/avtMassVoxelExtractor.C
@@ -2090,7 +2090,9 @@ avtMassVoxelExtractor::computePixelColor(double source_rgb[4], double dest_rgb[4
 
         // cos(angle) = a.b;  angle between normal and light
         float normal_dot_light = dot(gradient,dir);   // angle between light and normal;
-        normal_dot_light = std::max(0.0f, std::min(fabs(normal_dot_light),1.0f) );
+        normal_dot_light = std::max((double)0.0f,
+                                    std::min(fabs(normal_dot_light),
+                                             (double)1.0f) );
 
         // opacity correction
         float opacityCorrectiong = 0.7;
